organize code into packages
