--- src/main/java/Format.java
@@ -1,21 +1,0 @@
-/**
- * Provides the format that Duke's response will be displayed */
-public class Format {
-    public static final String SEPARATOR = "--------------------------------------\n";
-    public static void wrapPrint (String msg) {
-        String[] splitted = msg.split("\n");
-        String wrapee = "";
-        for (String line : splitted) {
-            wrapee = wrapee.concat("    " + line + "\n");
-        }
-        System.out.println(wrap(wrapee));
-    }
-
-    private static String wrap(String msg) {
-        return SEPARATOR + msg + SEPARATOR;
-    }
-
-    public static void printCursor() {
-        System.out.println(">>>");
-    }
-}
