--- src/main/java/NoTimeGivenException.java
@@ -1,5 +1,0 @@
-class NoTimeGivenException extends DukeException {
-    public NoTimeGivenException(String msg) {
-        super(msg);
-    }
-}
