+++ src/main/java/duke/Ui.java
@@ -1,3 +1,4 @@
+package duke;
 /**
  * Provides the format that Duke's response will be displayed */
 public class Ui {
