+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,4 @@
+package duke.command;
 public abstract class Command {
     public abstract void execute();
 }
