+++ src/main/java/duke/command/CommandBye.java
@@ -1,4 +1,10 @@
-class CommandBye extends Command {
+package duke.command;
+
+import duke.Over;
+import duke.Response;
+import duke.Ui;
+
+public class CommandBye extends Command {
     Over over;
     public CommandBye(Over over) {
         this.over = over;
