+++ src/main/java/duke/command/CommandList.java
@@ -1,4 +1,7 @@
-class CommandList extends Command {
+package duke.command;
+import duke.TaskList;
+import duke.Ui;
+public class CommandList extends Command {
     TaskList taskList;
 
     public CommandList(TaskList taskList) {
