+++ src/main/java/duke/command/CommandMark.java
@@ -1,4 +1,7 @@
-class CommandMark extends Command {
+package duke.command;
+import duke.TaskList;
+import duke.Ui;
+public class CommandMark extends Command {
     TaskList taskList;
     int taskNo;
