+++ src/main/java/duke/command/CommandUnclear.java
@@ -1,4 +1,6 @@
-class CommandUnclear extends Command {
+package duke.command;
+import duke.Ui;
+public class CommandUnclear extends Command {
     @Override
     public void execute() {
         Ui.wrapPrint("I'm sorry. I don't understand what that means :d");
