+++ src/main/java/duke/command/CommandUnmark.java
@@ -1,4 +1,7 @@
-class CommandUnmark extends Command {
+package duke.command;
+import duke.TaskList;
+import duke.Ui;
+public class CommandUnmark extends Command {
     TaskList taskList;
     int taskNo;
