+++ src/main/java/duke/dukeException/DukeException.java
@@ -1,4 +1,5 @@
-class DukeException extends Exception {
+package duke.dukeException;
+public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
     }
