+++ src/main/java/duke/dukeException/NoTimeGivenException.java
@@ -1,0 +1,6 @@
+package duke.dukeException;
+public class NoTimeGivenException extends DukeException {
+    public NoTimeGivenException(String msg) {
+        super(msg);
+    }
+}
